Possible fix for bad loads of saves Little fixes to stacks saving and special orb skills keeping track of orbs spawned

+++ Utopia/Classes/Battle/Skills/SkillBombs.m
@@ -68,7 +68,7 @@
   
   [self dropBombsOnPlayer:numOrbs];
   
-  [self showSkillPopupMiniOverlay:[NSString stringWithFormat:@"%li DMG", totalDamage]];
+  [self showSkillPopupMiniOverlay:[NSString stringWithFormat:@"%li DMG", (long)totalDamage]];
 
   
   return YES;
